2. На первой вкладке отобразил список друзей пользователя. 3. На второй вкладке отобразил фото выбраного друга
