--- newProject/newProject/010/GreenCollectionViewCell.swift
@@ -1,14 +1,0 @@
-//
-//  GreenCollectionViewCell.swift
-//  newProject
-//
-//  Created by serge on 10/03/2019.
-//  Copyright © 2019 serge. All rights reserved.
-//
-
-import UIKit
-
-class GreenCollectionViewCell: UICollectionViewCell {
-    @IBOutlet weak var cityLabel: UILabel!
-    
-}
