+++ newProject/newProject/Helpers/Extension.swift
@@ -1,0 +1,15 @@
+//
+//  Extension.swift
+//  newProject
+//
+//  Created by serge on 16/03/2019.
+//  Copyright © 2019 serge. All rights reserved.
+//
+
+import UIKit
+
+extension NSObject {
+    static func className() -> String {
+        return String(describing: self)
+    }
+}
