+++ newProject/newProject/Helpers/UITableViewExtension.swift
@@ -1,0 +1,36 @@
+//
+//  UITableViewExtension.swift
+//  newProject
+//
+//  Created by serge on 16/03/2019.
+//  Copyright © 2019 serge. All rights reserved.
+//
+
+import UIKit
+
+extension UITableView {
+    
+//    func registerNib(forClass neededClass: AnyClass) {
+//        register(UINib(nibName: String(describing: neededClass), bundle: Bundle.main), forCellReuseIdentifier: String(describing: neededClass))
+//    }
+//
+//    func registerCell(forClass neededClass: AnyClass) {
+//        register(neededClass, forCellReuseIdentifier: String(describing: neededClass))
+//    }
+//
+//    public func register(nib: UINib?, forHeaderFooterViewClass neededClass: AnyClass) {
+//        register(neededClass, forHeaderFooterViewReuseIdentifier: String(describing: neededClass))
+//    }
+//
+//    public func register(forHeaderFooterViewClass neededClass: AnyClass) {
+//        register(neededClass, forHeaderFooterViewReuseIdentifier: String(describing: neededClass))
+//    }
+//
+//    func dequeueReusableCell<T: UITableViewCell> (forClass neededClass: T.Type, for indexPath: IndexPath) -> T {
+//        return dequeueReusableCell(withIdentifier: String(describing: neededClass), for: indexPath) as! T
+//    }
+//
+//    public func dequeueReusableHeaderFooterView<T: UITableViewHeaderFooterView>(forClass neededClass: T.Type) -> T {
+//        return dequeueReusableHeaderFooterView(withIdentifier: String(describing: neededClass)) as! T
+//    }
+}
